End of Season 2022 ✨✡
Wow! What an incredible season we had :relaxed:
We learned so much and we are really proud of what we accomplished.

Simon we love you 💖❣ You taught us so much and we appreciate your help and efforts a lot.

To Asaf, you were a great a head of our lovely programming team (rashaz shshs) and we already miss you 😢
Thank you for making the iconic programming playlist and we are sure that we will come back to it the next season 🎶

We are incredibly excited for the next season and we can't wait.
Thank you to all of the people who took part in this amazing season 💕

Love you all,
Programming Team 2022

Co-Authored-By: Konata <42537566+konata-chan404@users.noreply.github.com>
Co-Authored-By: yoavkaplan123 <70645788+yoavkaplan123@users.noreply.github.com>
Co-Authored-By: noya-isaiah <67971686+noya-isaiah@users.noreply.github.com>
Co-Authored-By: Asaf Harel <45461162+Asaf-Harel@users.noreply.github.com>
Co-Authored-By: yuvalrader <45494935+yuvalrader@users.noreply.github.com>
Co-Authored-By: Simon Kharmatsky <22105518+ScarletMetal@users.noreply.github.com>